# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: API Tropa Livresca
  version: 1.0.0
  description: >-
    API do site Tropa Livresca. Rotas de cliente usam o cookie auth-token;
    rotas administrativas usam admin-token e também verificam is_admin. As
    rotas de usuários e movimentações administrativas exigem is_master.
    Erros usam {status, statusCode, error}; statusCode é o código HTTP enviado
    pelo middleware. Listas paginadas usam page (padrão 1), limit (padrão 12)
    e meta.totalItems/totalPages quando indicado.
servers:
  - url: /api/v1
    description: API versionada
...
paths:
...
  /clients/loja/venda/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Loja]
      summary: Consultar venda
      operationId: consultarVenda
      security:
        - clientCookieAuth: []
      responses:
        "201":
          $ref: "#/components/responses/Success"
        default:
          $ref: "#/components/responses/Error"
...
components:
  securitySchemes:
    clientCookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
      description: Cookie de sessão do usuário.
    adminCookieAuth:
      type: apiKey
      in: cookie
      name: admin-token
      description: Cookie de sessão administrativa.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Identificador no formato aceito pela rota (bigint ou UUID conforme o recurso).
...
  responses:
    Success:
      description: Requisição processada com sucesso.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
    Error:
      description: A requisição não pôde ser processada.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true